Add files via upload
The packages that are installed are specifically chosen for running Webull Desktop inside the container. And I really like nala. Only sudo and libcap2-bin are required for the toolbox to work. GUI apps may need the libraries being installed with nala, as they were dependencies for Webull. Not sure if you'd need the same ones for other GUI apps.

Note: when running apps inside the container, you'll need to create a script that sets the hostname to the same as the host OS, and then runs the application. Your .desktop file will execute that script to run the app. This will ensure that the window the app is open in is properly linked to the app icon.